Moved code into the oauth2 module.
